기능 업데이트

+++ src/main/java/com/goldyframework/repository/RepositoryServiceImpl.java
@@ -18,7 +18,6 @@ import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.FilenameUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.InputStreamResource;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.multipart.MultipartFile;
@@ -48,7 +47,7 @@ public class RepositoryServiceImpl implements RepositoryService {
 	/**
 	 * 저장소 바디
 	 */
-	private final RepositoryBody repository;
+	private final RepositoryBody body;
 	
 	/**
 	 * {@link RepositoryServiceImpl} 클래스의 새 인스턴스를 초기화 합니다.
@@ -57,10 +56,9 @@ public class RepositoryServiceImpl implements RepositoryService {
 	 * @param body
 	 *            저장소 바디
 	 */
-	@Autowired
 	public RepositoryServiceImpl(final RepositoryBody body) {
 		
-		this.repository = NullGtils.throwIfNull(body);
+		this.body = NullGtils.throwIfNull(body);
 	}
 	
 	/**
@@ -125,7 +123,7 @@ public class RepositoryServiceImpl implements RepositoryService {
 		final File file = this.getFile();
 		final String extension = FilenameUtils.getExtension(file.getAbsolutePath());
 		
-		final String fileName = this.repository.getDownloadName() + "." + extension;
+		final String fileName = this.body.getDownloadName() + "." + extension;
 		final String docName = new String(fileName.getBytes(Prop.DEFAULT_CHARSET), "ISO-8859-1");
 		final String contentDispositionHeader = "attachment; filename=\"" + docName + "\"";
 		final InputStreamResource body = new InputStreamResource(new FileInputStream(file));
@@ -147,7 +145,7 @@ public class RepositoryServiceImpl implements RepositoryService {
 	@Override
 	public File getFile() throws SQLException, RepositoryException, NotRegisteredFileException {
 		
-		return this.repository.getRegisteredFile();
+		return this.body.getRegisteredFile();
 	}
 	
 	/**
@@ -178,24 +176,22 @@ public class RepositoryServiceImpl implements RepositoryService {
 	 * {@inheritDoc}
 	 *
 	 * @author 2017. 6. 18. 오후 1:56:08 jeong
-	 * @throws NotRegisteredFileException
-	 * @throws RepositoryException
-	 * @throws SQLException
 	 */
 	@Override
-	public File save(final MultipartFile multipartFile)
-		throws IOException {
+	public File save(final MultipartFile multipartFile) throws IOException {
 		
 		ObjectInspection.checkNull(multipartFile);
 		
-		try {
-			this.delete();
-		} catch (SQLException | RepositoryException e) {
-			// do nothig;
+		if (this.body.getNamingType().isRemoveIfDuplication()) {
+			try {
+				this.delete();
+			} catch (SQLException | RepositoryException e) {
+				// do nothig;
+			}
 		}
 		
 		final String extension = FilenameUtils.getExtension(multipartFile.getOriginalFilename());
-		final File file = this.repository.generateSavePath(extension);
+		final File file = this.body.generateSavePath(extension);
 		
 		final File directory = file.getParentFile();
 		if (directory.exists() == false) {
@@ -222,7 +218,7 @@ public class RepositoryServiceImpl implements RepositoryService {
 			file = this.getFile();
 		} catch (final NotRegisteredFileException e) {
 			LOGGER.trace("파일이 없어 신규 파일을 생성합니다.", e);
-			file = this.repository.generateSavePath();
+			file = this.body.generateSavePath();
 		}
 		final boolean success = file.createNewFile();
 		SonarHelper.noStatic(success);
